+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.AI.API/Models/ApiDtos/ArticleApiDto.cs
@@ -1,4 +1,4 @@
-namespace FF.Articles.Backend.Common.ApiDtos;
+namespace FF.Articles.Backend.AI.API.Models.ApiDtos;
 public class ArticleApiDto
 {
     public long ArticleId { get; set; }
@@ -8,9 +8,6 @@ public class ArticleApiDto
     public string Title { get; set; }
     public string Content { get; set; }
     public string Abstract { get; set; }
-    /// <summary>
-    /// 1: article, 2: sub article, 3: topic article
-    /// </summary>
     public string ArticleType { get; set; }
     public long? ParentArticleId { get; set; }
     public List<ArticleApiDto> SubArticles { get; set; } = new();
@@ -22,5 +19,11 @@ public class ArticleApiDto
     public int SortNumber { get; set; }
     public int IsHidden { get; set; }
 
+}
+public static class ArticleTypes
+{
+    public const string Article = "Article";
+    public const string SubArticle = "SubArticle";
+    public const string TopicArticle = "TopicArticle";
 }