+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Constants/ArticleTypes.cs
@@ -2,14 +2,7 @@ namespace FF.Articles.Backend.Contents.API.Constants;
 
 public static class ArticleTypes
 {
-    /// <summary>
-    /// Default
-    /// </summary>
     public const string Article = "Article";
-    /// <summary>
-    /// SubArticle
-    /// </summary>
     public const string SubArticle = "SubArticle";
-
     public const string TopicArticle = "TopicArticle";
 }