+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Requests/Topics/TopicAddRequest.cs
@@ -6,7 +6,7 @@ public class TopicAddRequest
     public string Abstract { get; set; } = string.Empty;
     //public string Content { get; set; } = string.Empty;
     public string TopicImage { get; set; } = string.Empty;
-    public string Category { get; set; } = TopicCategories.Other;
+    public string Category { get; set; } = "Other";
     public int SortNumber { get; set; } = 1;
     public int IsHidden { get; set; } = 0;
 }