+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
     - P2: Add quick article generation
     - P2: Generate Topic Article
     - P2: Generate existing topic
+    - P2: Regarnerate Article
     - P2: Article add page
     - P4: Fix Ai Gen possible codefence
     - P4: Batch update article/topic/tag